backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
openai
python-jose
passlib[bcrypt]
pinecone
azure-ai-formrecognizer
azure-core
duckduckgo-search
python-multipart
psycopg2-binary
pydantic[email]